🎨 Move models to separate files
