+++ Sources/ACKLocalizationCore/Model/PluralTranslationKey.swift
@@ -1,0 +1,17 @@
+//
+//  PluralTranslationKey.swift
+//  
+//
+//  Created by Lukáš Hromadník on 07/07/2020.
+//
+
+import Foundation
+
+enum PluralTranslationKey: String, Codable {
+    case zero
+    case one
+    case two
+    case few
+    case many
+    case other
+}
